--- change-beta/@azure-communication-react-3130dff6-1b39-4f6f-a2fe-5088f25d61b2.json
@@ -1,9 +1,0 @@
-{
-  "type": "minor",
-  "area": "feature",
-  "workstream": "Storybook V8 Migration",
-  "comment": "Update storybook files for SendBox",
-  "packageName": "@azure/communication-react",
-  "email": "longamy@microsoft.com",
-  "dependentChangeType": "patch"
-}
