+++ packages/storybook8/stories/Components/SendBox/Docs.mdx
@@ -16,7 +16,7 @@ import MentionsExampleText from '!!raw-loader!./snippets/Mentions.snippet.tsx';
 import SendBoxExampleText from '!!raw-loader!./snippets/SendBox.snippet.tsx';
 import SendBoxWithSystemMessageExampleText from '!!raw-loader!./snippets/SendBoxWithSystemMessage.snippet.tsx';
 
-import * as SendBoxStories from './SendBox.stories';
+import * as SendBoxStories from './index.stories';
 
 
 <Meta of={SendBoxStories} component={SendBox} />
