--- packages/storybook8/stories/Components/SendBox/SendBox.stories.tsx
@@ -1,101 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { SendBox as SendBoxComponent } from '@azure/communication-react';
-import { Meta } from '@storybook/react';
-import React, { useEffect, useRef } from 'react';
-import { controlsToAdd, hiddenControl } from '../../controlsUtils';
-
-import { AttachmentUploadsExample } from './snippets/AttachmentUploads.snippet';
-import { CustomIconExample } from './snippets/CustomIcon.snippet';
-import { CustomStylingExample } from './snippets/CustomStyling.snippet';
-import { MentionsExample } from './snippets/Mentions.snippet';
-import { SendBoxExample } from './snippets/SendBox.snippet';
-import { SendBoxWithSystemMessageExample } from './snippets/SendBoxWithSystemMessage.snippet';
-
-const Default = (args): JSX.Element => {
-  const timeoutRef = useRef<NodeJS.Timeout>();
-  const delayForSendButton = 300;
-
-  useEffect(() => {
-    // Clean up the timeout when the component unmounts
-    return () => {
-      if (timeoutRef.current) clearTimeout(timeoutRef.current);
-    };
-  }, []);
-
-  return (
-    <div style={{ width: '31.25rem' }}>
-      <SendBoxComponent
-        {...args}
-        onSendMessage={async (message, options) => {
-          timeoutRef.current = setTimeout(() => {
-            alert(`Sent message: "${message}" with options: ${JSON.stringify(options)}`);
-          }, delayForSendButton);
-        }}
-        onTyping={() => {
-          console.log(`Sending typing notifications`);
-          return Promise.resolve();
-        }}
-        systemMessage={args.hasWarning ? args.warningMessage : undefined}
-        attachments={
-          args.hasAttachments
-            ? [{ id: 'f2d1fce73c98', name: 'file1.txt', progress: 1, url: 'https://www.bing.com' }]
-            : undefined
-        }
-        onCancelAttachmentUpload={(attachmentId) => {
-          window.alert(`onCancelAttachmentUpload callback is called for attachment id: "${attachmentId}"`);
-        }}
-      />
-    </div>
-  );
-};
-
-Default.storyName = 'SendBox';
-
-export default {
-  title: 'Components/SendBox/Send Box',
-  component: SendBoxComponent,
-  argTypes: {
-    attachments: hiddenControl,
-    onCancelAttachmentUpload: hiddenControl,
-    onRenderAttachmentUploads: hiddenControl,
-    disabled: controlsToAdd.disabled,
-    hasWarning: controlsToAdd.isSendBoxWithWarning,
-    warningMessage: controlsToAdd.sendBoxWarningMessage,
-    hasAttachments: controlsToAdd.isSendBoxWithAttachments,
-    systemMessage: hiddenControl,
-    onSendMessage: hiddenControl,
-    onTyping: hiddenControl,
-    onRenderSystemMessage: hiddenControl,
-    supportNewline: hiddenControl,
-    onRenderIcon: hiddenControl,
-    styles: hiddenControl,
-    strings: hiddenControl
-  }
-} as Meta<typeof SendBoxComponent>;
-
-export const SendBox = Default.bind({});
-
-export const AttachmentUploadsExampleDocsOnly = {
-  render: AttachmentUploadsExample
-};
-export const CustomIconExampleDocsOnly = {
-  render: CustomIconExample
-};
-
-export const CustomStylingExampleDocsOnly = {
-  render: CustomStylingExample
-};
-
-export const MentionsExampleDocsOnly = {
-  render: MentionsExample
-};
-
-export const SendBoxExampleDocsOnly = {
-  render: SendBoxExample
-};
-
-export const SendBoxWithSystemMessageExampleDocsOnly = {
-  render: SendBoxWithSystemMessageExample
-};
