+++ packages/storybook8/stories/Components/SendBox/snippets/SendBox.snippet.tsx
@@ -1,5 +1,9 @@
-import { SendBox, FluentThemeProvider } from '@azure/communication-react';
+import { SendBox, FluentThemeProvider, DEFAULT_COMPONENT_ICONS } from '@azure/communication-react';
+import { initializeIcons, registerIcons } from '@fluentui/react';
 import React from 'react';
+
+initializeIcons();
+registerIcons({ icons: DEFAULT_COMPONENT_ICONS });
 
 export const SendBoxExample: () => JSX.Element = () => (
   <FluentThemeProvider>
